Enabled by default import_notes_ignore_deck_movement to
Do Not Move Existing Cards: By default on import of a CrowdAnki deck
file, when a note already exists in Anki itself, all notes will be
updated and placed in the deck set in the deck file. Tick this box if
 you wish only to have the notes updated, but left in their current
deck. See <a href=\"https://github.com/Stvad/CrowdAnki/issues/23